Move contexts/StreamrClient to components/StreamrClientProvider
