+++ app/src/editor/dashboard/index.jsx
@@ -9,7 +9,7 @@ import LoadingIndicator from '$shared/components/LoadingIndicator'
 import ErrorComponentView from '$shared/components/ErrorComponentView'
 import UndoControls from '$editor/shared/components/UndoControls'
 import { Context as UndoContext, Provider as UndoContextProvider } from '$shared/contexts/Undo'
-import ClientProvider from '$shared/contexts/StreamrClient'
+import ClientProvider from '$shared/components/StreamrClientProvider'
 import * as sharedServices from '$editor/shared/services'
 import { SelectionProvider } from '$shared/hooks/useSelection'
 import { Provider as PendingProvider } from '$shared/contexts/Pending'
