+++ app/src/editor/shared/tests/ModuleSubscription.test.jsx
@@ -4,7 +4,7 @@ import { Provider } from 'react-redux'
 import { useClient } from 'streamr-client-react'
 import mockStore from '$testUtils/mockStoreProvider'
 import { setupAuthorizationHeader, loadModuleDefinition } from '$editor/shared/tests/utils'
-import ClientProvider from '$shared/contexts/StreamrClient'
+import ClientProvider from '$shared/components/StreamrClientProvider'
 import ModuleSubscription from '$editor/shared/components/ModuleSubscription'
 import * as State from '$editor/canvas/state'
 import * as Services from '$editor/canvas/services'
