+++ app/src/marketplace/containers/StreamPreviewPage/index.jsx
@@ -8,7 +8,7 @@ import { I18n } from 'react-redux-i18n'
 import { useClient } from 'streamr-client-react'
 import { Context as RouterContext } from '$shared/contexts/Router'
 import ProductController, { useController } from '../ProductController'
-import ClientProvider from '$shared/contexts/StreamrClient'
+import ClientProvider from '$shared/components/StreamrClientProvider'
 import Subscription from '$shared/components/Subscription'
 import { Provider as SubscriptionStatusProvider } from '$shared/contexts/SubscriptionStatus'
 import usePending from '$shared/hooks/usePending'
