+++ app/src/shared/components/ActivityList/ActivityStreamHandler.jsx
@@ -2,7 +2,7 @@
 
 import React, { useEffect, useState } from 'react'
 import { useClient } from 'streamr-client-react'
-import ClientProvider from '$shared/contexts/StreamrClient'
+import ClientProvider from '$shared/components/StreamrClientProvider'
 import Activity from '$shared/utils/Activity'
 import { isLocalStorageAvailable } from '$shared/utils/storage'
 import { Provider as PendingProvider } from '$shared/contexts/Pending'
