+++ app/src/shared/components/ActivityList/index.jsx
@@ -3,7 +3,7 @@ import { useSelector } from 'react-redux'
 import { selectPendingTransactions } from '$mp/modules/transactions/selectors'
 import Activity, { actionTypes } from '$shared/utils/Activity'
 import { isLocalStorageAvailable } from '$shared/utils/storage'
-import ClientProvider from '$shared/contexts/StreamrClient'
+import ClientProvider from '$shared/components/StreamrClientProvider'
 import Subscription from '$shared/components/Subscription'
 import { useFetchResource } from './ActivityResourceProvider'
 import Items from './Items'
