+++ app/src/shared/components/StreamrClientProvider.jsx
