+++ app/src/userpages/components/StreamPage/Edit/PreviewView.jsx
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 import { Translate } from 'react-redux-i18n'
 import { useClient } from 'streamr-client-react'
 import Button from '$shared/components/Button'
-import ClientProvider from '$shared/contexts/StreamrClient'
+import ClientProvider from '$shared/components/StreamrClientProvider'
 import useModal from '$shared/hooks/useModal'
 import Subscription from '$shared/components/Subscription'
 import { Provider as SubscriptionStatusProvider } from '$shared/contexts/SubscriptionStatus'
